requirements.txt
numpy>=1.21.0,<1.25.0
pandas>=1.5.0,<2.1.0
streamlit>=1.25.0
plotly>=5.10.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.11.0
pillow>=9.0.0
openpyxl>=3.0.0
reportlab>=3.6.0

